requirements.txt
fastapi
uvicorn
openai
python-dotenv
numpy
scikit-learn
